Refact: map 동작 함수 파일 분리
makeNewMap, addressToCoords, makeNewMarker -> MapFunction.js
